\name{breastcancer}
\docType{data}
\alias{breastcancer}
\title{Breast Cancer Identification}
\description{
  Physical measurements from breast masses for distinguishing malignant from benign cases.
}
\usage{breastcancer}
\format{A list with entries $x (a 699x9 matrix with each row corresponding to an individual case) and $c (a vector of labels, benign = 2 and malignant = 4.}
\source{UCI Machine Learning Repository.}
\references{
  Dheeru, D. and E. Karra Taniskidou (2017). UCI machine learning repository. \url{https://archive.ics.uci.edu/ml}
  
  Wolberg, W.H., & Mangasarian, O.L. (1990). Multisurface method of pattern separation for medical diagnosis applied to breast cytology. In Proceedings of the National Academy of Sciences, 87, 9193--9196.

Zhang, J. (1992). Selecting typical instances in instance-based learning. In Proceedings of the Ninth International Machine Learning Conference (pp. 470--479). Aberdeen, Scotland: Morgan Kaufmann.
}
\keyword{datasets}
