\name{controlchart}
\docType{data}
\alias{controlchart}
\title{Simulated time-series data.}
\description{
  Synthetically generated control charts. There are six different classes:

1. Normal
2. Cyclic
3. Increasing trend
4. Decreasing trend
5. Upward shift
6. Downward shift 
}
\usage{controlchart}
\format{A list with entries $x (a 600x60 matrix with each row corresponding to an individual time series) and $c (a vector of labels)}
\source{UCI Machine Learning Repository.}
\references{
  Dheeru, D. and E. Karra Taniskidou (2017). UCI machine learning repository. \url{https://archive.ics.uci.edu/ml}
  
  Alcock R.J. and Manolopoulos Y. Time-Series Similarity Queries Employing a Feature-Based Approach. 7th Hellenic Conference on Informatics. August 27-29. Ioannina,Greece 1999.

  D.T. Pham and A.B. Chan "Control Chart Pattern Recognition using a New Type of Self Organizing Neural Network" Proc. Instn, Mech, Engrs. Vol 212, No 1, pp 115-127 1998.
}
\keyword{datasets}
