\name{glass}
\docType{data}
\alias{glass}
\title{Glass identification data set.}
\description{
  Identification of glass source based on physical and chemical properties. There are seven different classes, six of which are present in this data set.
}
\usage{glass}
\format{A list with entries $x (a 214x9 matrix with each row corresponding to an individual observation) and $c (a vector of labels)}
\source{UCI Machine Learning Repository.}
\references{
  Dheeru, D. and E. Karra Taniskidou (2017). UCI machine learning repository. \url{https://archive.ics.uci.edu/ml}
  
  Ian W. Evett and Ernest J. Spiehler. Rule Induction in Forensic Science. Central Research Establishment. Home Office Forensic Science Service. Aldermaston, Reading, Berkshire RG7 4PN
}
\keyword{datasets}
