\name{iris}
\docType{data}
\alias{iris}
\title{Iris data set.}
\description{
  Physical measurements from 3 species of iris plants.
}
\usage{iris}
\format{A list with entries $x (a 150x4 matrix with each row corresponding to an individual plant) and $c (a vector of labels)}
\source{UCI Machine Learning Repository.}
\references{
  Dheeru, D. and E. Karra Taniskidou (2017). UCI machine learning repository. \url{https://archive.ics.uci.edu/ml}
  
  Fisher,R.A. "The use of multiple measurements in taxonomic problems" Annual Eugenics, 7, Part II, 179-188 (1936); also in "Contributions to Mathematical Statistics" (John Wiley, NY, 1950).
}
\keyword{datasets}
