\name{libras}
\docType{data}
\alias{libras}
\title{Libras hand movement data set.}
\description{
  The data contain coordinate features derived from 360 videos of hand movements. Each instance describes one of 15 different hand movement types in LIBRAS.
}
\usage{libras}
\format{A list with entries $x (a 360x90 matrix with each row corresponding to an individual hand movement) and $c (a vector of labels)}
\source{UCI Machine Learning Repository.}
\references{
  Dheeru, D. and E. Karra Taniskidou (2017). UCI machine learning repository. \url{https://archive.ics.uci.edu/ml}
  
  DIAS, D. B.; MADEO, R. C. B.; ROCHA, T.; BÍSCARO, H. H.; PERES, S. M..
Hand Movement Recognition for Brazilian Sign Language: A Study Using Distance-Based Neural Networks.
In: 2009 International Joint Conference on Neural Networks, 2009, Atlanta, GA.
}
\keyword{datasets}
