\name{mammography}
\docType{data}
\alias{mammography}
\title{Mammographic mass data set.}
\description{
  The data contain five attributes describing mammographic mass cases. One attribute is non-predictive, the others are age of patient and three simple physical characteristics of the masses. Classes are benign = 0 and malignant = 1.
}
\usage{mammography}
\format{A list with entries $x (a 961x5 matrix with each row corresponding to an individual case) and $c (a vector of labels)}
\source{UCI Machine Learning Repository.}
\references{
  Dheeru, D. and E. Karra Taniskidou (2017). UCI machine learning repository. \url{https://archive.ics.uci.edu/ml}
  
  M. Elter, R. Schulz-Wendtland and T. Wittenberg (2007)
The prediction of breast cancer biopsy outcomes using two CAD approaches that both emphasize an intelligible decision process.
Medical Physics 34(11), pp. 4164-4172
}
\keyword{datasets}
