\name{mfdigits}
\docType{data}
\alias{mfdigits}
\title{Multiple feature digits data set.}
\description{
  This dataset consists of features derived from handwritten digits (0--9) taken from Dutch utility maps. We have included only the 210 profile correlation features.
}
\usage{mfdigits}
\format{A list with entries $x (a 2000x210 matrix with each row corresponding to an individual handwritten digit) and $c (a vector of labels)}
\source{UCI Machine Learning Repository.}
\references{
  Dheeru, D. and E. Karra Taniskidou (2017). UCI machine learning repository. \url{https://archive.ics.uci.edu/ml}
  
  M. van Breukelen, R.P.W. Duin, D.M.J. Tax, and J.E. den Hartog, Handwritten digit recognition by combined classifiers, Kybernetika, vol. 34, no. 4, 1998, 381-386.
}
\keyword{datasets}
