\name{oliveoil}
\docType{data}
\alias{oliveoil}
\title{Olive oil data.}
\description{
  Chemical composition data of 572 olive oils. Two label sets, one describing the macro-area (one of three) from which the data were taken, and one describing the region (one of nine). This data set was taken from the CRAN package 'pdfCluster'.
}
\usage{oliveoil}
\format{A list with entries $x (a 572x8 matrix with each row corresponding to an oil instance), $c1 (a vector of macro-area labels), and $c2 (a vector of region labels)}
\source{Forina, M., Lanteri, S. Armanino, C., Casolino, C., Casale, M., Oliveri, P. (2008). V-PARVUS. An Extendible Package of programs for explorative data analysis, classification and regression analysis. Dip. Chimica e Tecnologie Farmaceutiche ed Alimentari, Università di Genova. Freely available at http://www.parvus.unige.it.

CRAN package 'pdfCluster'.
}
\references{
  Azzalini, A. and G. Menardi (2014). Clustering via nonparametric density estimation: the
R package pdfCluster. Journal of Statistical Software 57 (11), 1–26.
}
\keyword{datasets}
