\name{optidigits}
\docType{data}
\alias{optidigits}
\title{Optical recognition of handwritten digits data set.}
\description{
  This dataset consists vectorised images of handwritten digits (0--9). The images have been compressed to 8x8, resulting in 64 dimensional observations.
}
\usage{optidigits}
\format{A list with entries $x (a 5620x64 matrix with each row corresponding to an individual image) and $c (a vector of labels)}
\source{UCI Machine Learning Repository.}
\references{
  Dheeru, D. and E. Karra Taniskidou (2017). UCI machine learning repository. \url{https://archive.ics.uci.edu/ml}
  
  C. Kaynak (1995) Methods of Combining Multiple Classifiers and Their Applications to Handwritten Digit Recognition, MSc Thesis, Institute of Graduate Studies in Science and Engineering, Bogazici University.

E. Alpaydin, C. Kaynak (1998) Cascading Classifiers, Kybernetika.
}
\keyword{datasets}
