\name{parkinsons}
\docType{data}
\alias{parkinsons}
\title{Parkinsons data set.}
\description{
  This dataset contains biomedical voice measurements from 31 individuals, 23 of whom have Parkinson's disease (PD). Classes correspond to healthy = 0 and with PD = 1.
}
\usage{parkinsons}
\format{A list with entries $x (a 195x22 matrix with each row corresponding to an individual case) and $c (a vector of labels)}
\source{UCI Machine Learning Repository.}
\references{
  Dheeru, D. and E. Karra Taniskidou (2017). UCI machine learning repository. \url{https://archive.ics.uci.edu/ml}
  
  Max A. Little, Patrick E. McSharry, Eric J. Hunter, Lorraine O. Ramig (2008), 'Suitability of dysphonia measurements for telemonitoring of Parkinson's disease', IEEE Transactions on Biomedical Engineering.
}
\keyword{datasets}
