\name{pendigits}
\docType{data}
\alias{pendigits}
\title{Pen-based recognition of handwritten digits data set.}
\description{
  This dataset contains coordinate information from hand trajectories during the writing of one of the digits 0--9. We include only the training data.
}
\usage{pendigits}
\format{A list with entries $x (a 7494x16 matrix with each row corresponding to an individual handwritten digit) and $c (a vector of labels)}
\source{UCI Machine Learning Repository.}
\references{
  Dheeru, D. and E. Karra Taniskidou (2017). UCI machine learning repository. \url{https://archive.ics.uci.edu/ml}
  
  F. Alimoglu (1996) Combining Multiple Classifiers for Pen-Based Handwritten Digit Recognition, MSc Thesis, Institute of Graduate Studies in Science and Engineering, Bogazici University.

F. Alimoglu, E. Alpaydin, "Methods of Combining Multiple Classifiers Based on Different Representations for Pen-based Handwriting Recognition," Proceedings of the Fifth Turkish Artificial Intelligence and Artificial Neural Networks Symposium (TAINN 96), June 1996, Istanbul, Turkey.
}
\keyword{datasets}
