\name{phoneme}
\docType{data}
\alias{phoneme}
\title{Phoneme data set.}
\description{
  This data set contains features extracted from spoken utterances of five different vocal sounds common in speech.
}
\usage{phoneme}
\format{A list with entries $x (a 4509x256 matrix with each row corresponding to an individual utterance) and $c (a vector of labels)}
\source{The Elements of Statistical Learning.}
\references{
  Friedman, J., T. Hastie, and R. Tibshirani (2001). The Elements of Statistical Learning,
Volume 1. Springer series in statistics New York, NY, USA.
}
\keyword{datasets}
