\name{seeds}
\docType{data}
\alias{seeds}
\title{Wheat seed kernel data set.}
\description{
  This data set contains measurements of geometrical properties of kernels belonging to three different varieties of wheat.
}
\usage{seeds}
\format{A list with entries $x (a 210x7 matrix with each row corresponding to an individual seed) and $c (a vector of labels)}
\source{UCI Machine Learning Repository.}
\references{
  Dheeru, D. and E. Karra Taniskidou (2017). UCI machine learning repository. \url{https://archive.ics.uci.edu/ml}
  
  M. Charytanowicz, J. Niewczas, P. Kulczycki, P.A. Kowalski, S. Lukasik, S. Zak, 'A Complete Gradient Clustering Algorithm for Features Analysis of X-ray Images', in: Information Technologies in Biomedicine, Ewa Pietka, Jacek Kawa (eds.), Springer-Verlag, Berlin-Heidelberg, 2010, pp. 15-24.
}
\keyword{datasets}
